python/requirements.txt
autopep8==1.5
azure-functions==1.2.0
certifi==2019.11.28
chardet==3.0.4
emoji==0.5.4
idna==2.9
pycodestyle==2.5.0
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
six==1.14.0
twilio==6.35.5
urllib3==1.25.8
